tests: Move ransomware_target() fixture to ransomware/conftest.py
