+++ monkey/tests/unit_tests/infection_monkey/ransomware/test_ransomware_payload.py
@@ -1,6 +1,4 @@
 import os
-import shutil
-from pathlib import Path
 
 import pytest
 from tests.utils import hash_file, is_user_admin
@@ -34,15 +32,6 @@ TEST_KEYBOARD_TXT_ENCRYPTED_SHA256 = (
 
 def with_extension(filename):
     return f"{filename}{EXTENSION}"
-
-
-@pytest.fixture
-def ransomware_target(tmp_path, data_for_tests_dir):
-    ransomware_test_data = Path(data_for_tests_dir) / "ransomware_targets"
-    ransomware_target = tmp_path / "ransomware_target"
-    shutil.copytree(ransomware_test_data, ransomware_target)
-
-    return ransomware_target
 
 
 @pytest.fixture
